Rename: 커뮤니티 서비스 로직 이동
커뮤니티에서 사용되는 서비스 로직을 이동하였다. 다른 파트에서 새로운 기능이나 수정이 필요하다면 해당 클래스를 상속받아서 사용하면 된다.

Related to: #37
