+++ src/main/java/pulleydoreurae/careerquestbackend/common/community/service/CommentService.java
@@ -1,4 +1,4 @@
-package pulleydoreurae.careerquestbackend.basiccommunity.service;
+package pulleydoreurae.careerquestbackend.common.community.service;
 
 import java.util.List;
