+++ src/test/java/pulleydoreurae/careerquestbackend/basiccommunity/service/PostLikeServiceTest.java
@@ -27,6 +27,8 @@ import pulleydoreurae.careerquestbackend.common.community.exception.PostNotFound
 import pulleydoreurae.careerquestbackend.common.community.repository.PostLikeRepository;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.common.community.domain.entity.PostLike;
+import pulleydoreurae.careerquestbackend.common.community.service.CommonCommunityService;
+import pulleydoreurae.careerquestbackend.common.community.service.PostLikeService;
 
 /**
  * @author : parkjihyeok
